Improve error handling
Error handling of the faucet is improved considerably. In my local
tests, the process never crashed and there were no more warnings of
uncaught exceptions.

When the websocket connection of web3 breaks, the server immediately
returns an error to the client, but stays active. As soon as web3/the
websocket can reconnect to the node, the faucet is functional again.

I updated the error messaging in a way where the faucet would always log
details of the error, but it would not always return error details to
the client to not leak any potentially sensitive information to random
clients.

I updated the Logger to no longer handle 'uncaughtEvents', as that
interfered with the correct error handling.

I had to update the Web3js version (to `beta.52`) in the faucet, as the
old version we usually use (`beta.37`) doesn't do error handling correct.
It 'throws' in an event emitter and that cannot be caught.

With the updated Web3js it was no longer possible to use the
`ZeroClientProvider` together with the `Accounts` how we used to.
Furthermore, the latest beta automatically signs transactions locally,
if the `from` address is unlocked in `eth.accounts.wallet`.
So I removed the `ZeroClientProvider` and switched to the Web3js wallet
for accounts and signing.

The problem now is that Web3js doesn't do proper nonce handling. Thus,
I  added a very simple nonce handler to get started (`n + 1`). It
fetches the current nonce from the node only when it starts.

It also doesn't do gas estimation, so that needs to be done separately
now before sending a transaction.

There was also an issue that I cannot do gas estimation asynchronously
while still being able to access the `PromiEvent`'s events. When waiting
for a promise to resolve, it would also resolve the wrapped PromiEvent.
As a fix, I wrapped the `fill` responses in Promises which resolve (or
reject) upon the transaction hash (or error) event of the underlying
transaction. This way, we can do async/await calls to get the nonce or
the estimated gas of a transaction.

Also:

* I moved the ABI into a TypeScript class.
* Added static typing wherever I found it was missing.
* Generally refactored the code to be more in-line with best practices.

Fixes #9